Cleanup ReDB, remove KVStorage interface
