+++ src/storage/prelude.rs
@@ -1,6 +1,3 @@
-use redb::{AccessGuard, TypeName};
-use serde::{de::Deserialize, Serialize};
-use std::any::type_name;
 use std::fmt::Debug;
 use thiserror::Error;
 
@@ -58,94 +55,5 @@ pub enum CAStorageError {
 impl<T> From<std::sync::PoisonError<T>> for CAStorageError {
     fn from(_: std::sync::PoisonError<T>) -> Self {
         CAStorageError::PoisonedLock
-    }
-}
-
-pub use redb::{Key, Value};
-pub type Table<K, V> = redb::TableDefinition<'static, K, V>;
-
-// Trait for normal KV-Storage. Each key can be updated and store any data.
-// You can wrap this higher level for type-safe access.
-// This is kind of ugly now, fighting with ReDB types, so only works for ReDB now.
-// It would be nice to extend this to be valid for other storage backends later.
-pub trait KVStorage: Send + Sync {
-    /// Stores the given data and returns the digest to look it up later.
-    /// If the data was already stored, this is a no-op but still returns the digest with no error.
-    fn set<K: Key, V: Value>(
-        &self,
-        table: Table<K, V>,
-        key: K::SelfType<'_>,
-        value: &V::SelfType<'_>,
-    ) -> Result<(), KVStorageError>;
-
-    /// Looks up the data for a given digest and returns it. If data not present, returns CAStorageError::NotFound(_)
-    fn get<K: Key, V: Value + 'static>(
-        &self,
-        table: Table<K, V>,
-        key: K::SelfType<'_>,
-    ) -> Result<Option<AccessGuard<'static, V>>, KVStorageError>;
-
-    fn remove<K: Key, V: Value + 'static>(
-        &self,
-        table: Table<K, V>,
-        key: K::SelfType<'_>,
-    ) -> Result<(), KVStorageError>;
-
-    // TODO: add range
-}
-
-#[derive(Debug, Error)]
-pub enum KVStorageError {
-    // TODO: remove this later, make more generic
-    #[error("ReDB error: {0}")]
-    ReDB(redb::Error),
-}
-
-impl<T> From<T> for KVStorageError
-where
-    T: Into<redb::Error>,
-{
-    fn from(e: T) -> Self {
-        KVStorageError::ReDB(e.into())
-    }
-}
-
-/// Wrapper type to handle keys and values using bincode serialization
-#[derive(Debug, Clone)]
-pub struct JSON<T>(pub T);
-
-impl<T> Value for JSON<T>
-where
-    T: Debug + Serialize + for<'a> Deserialize<'a>,
-{
-    type SelfType<'a> = T
-    where
-        Self: 'a;
-
-    type AsBytes<'a> = Vec<u8>
-    where
-        Self: 'a;
-
-    fn fixed_width() -> Option<usize> {
-        None
-    }
-
-    fn from_bytes<'a>(data: &'a [u8]) -> Self::SelfType<'a>
-    where
-        Self: 'a,
-    {
-        serde_json::from_slice(data).unwrap()
-    }
-
-    fn as_bytes<'a, 'b: 'a>(value: &'a Self::SelfType<'b>) -> Self::AsBytes<'a>
-    where
-        Self: 'a,
-        Self: 'b,
-    {
-        serde_json::to_vec(value).unwrap()
-    }
-
-    fn type_name() -> TypeName {
-        TypeName::new(&format!("JSON<{}>", type_name::<T>()))
     }
 }
